Change functions for classes
