+++ src/components/abstractComponent.js
@@ -1,0 +1,27 @@
+import {createElement} from '../utils.js';
+
+export default class AbstractComponent {
+  constructor() {
+    this._element = null;
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    if (this._element) {
+      this._element = null;
+    }
+
+    return this._element;
+  }
+
+  getTemplate() {
+    throw Error(`Abstract method not implemented`);
+  }
+}
