+++ src/components/board-tasks.js
@@ -1,0 +1,7 @@
+import AbstractComponent from './abstractComponent.js';
+
+export default class BoardTasks extends AbstractComponent {
+  getTemplate() {
+    return `<div class="board__tasks"></div>`;
+  }
+}
