+++ src/components/board.js
@@ -1,9 +1,9 @@
-export const getBoard = () => {
-  return `<section class="board container">
-  <div class="board__filter-list">
-    <a href="#" class="board__filter">SORT BY DEFAULT</a>
-    <a href="#" class="board__filter">SORT BY DATE up</a>
-    <a href="#" class="board__filter">SORT BY DATE down</a>
-  </div>
-  </section>`;
-};
+import AbstractComponent from './abstractComponent.js';
+
+export default class Board extends AbstractComponent {
+  getTemplate() {
+    return `<section class="board container">
+    <div class="board__tasks"></div>
+    </section>`;
+  }
+}
