--- src/components/boardTasks.js
@@ -1,3 +1,0 @@
-export const getBoardTasks = () => {
-  return `<div class="board__tasks"></div>`;
-};
