+++ src/components/filter-section.js
@@ -1,0 +1,7 @@
+import AbstractComponent from './abstractComponent.js';
+
+export default class FilterSection extends AbstractComponent {
+  getTemplate() {
+    return `<section class="main__filter filter container"></section>`;
+  }
+}
