--- src/components/filterSection.js
@@ -1,3 +1,0 @@
-export const getFilterSection = () => {
-  return `<section class="main__filter filter container"></section>`;
-};
