+++ src/components/load-more.js
@@ -1,0 +1,7 @@
+import AbstractComponent from './abstractComponent.js';
+
+export default class LoadMore extends AbstractComponent {
+  getTemplate() {
+    return `<button class="load-more" type="button">load more</button>`;
+  }
+}
