--- src/components/loadMore.js
@@ -1,3 +1,0 @@
-export const getLoadMore = () => {
-  return `<button class="load-more" type="button">load more</button>`;
-};
