+++ src/components/task-card.js
@@ -1,0 +1,59 @@
+export const makeTask = ({description, dueDate, hours, minutes, tags, color, repeatingDays}) => {
+  return `<article class="card card--${color} ${Object.values(repeatingDays).some((elem) => elem) ? `card--repeat` : ``}">
+    <div class="card__form">
+      <div class="card__inner">
+        <div class="card__control">
+          <button type="button" class="card__btn card__btn--edit">
+            edit
+          </button>
+          <button type="button" class="card__btn card__btn--archive">
+            archive
+          </button>
+          <button
+            type="button"
+            class="card__btn card__btn--favorites card__btn--disabled"
+          >
+            favorites
+          </button>
+        </div>
+
+        <div class="card__color-bar">
+          <svg class="card__color-bar-wave" width="100%" height="10">
+            <use xlink:href="#wave"></use>
+          </svg>
+        </div>
+
+        <div class="card__textarea-wrap">
+          <p class="card__text">${description}</p>
+        </div>
+
+        <div class="card__settings">
+          <div class="card__details">
+            <div class="card__dates">
+              <div class="card__date-deadline">
+                <p class="card__input-deadline-wrap">
+                  <span class="card__date">${new Date(dueDate).toDateString()}</span>
+                  <span class="card__time">${hours}:${minutes} PM</span>
+                </p>
+              </div>
+            </div>
+
+            <div class="card__hashtag">
+              <div class="card__hashtag-list">
+              <span class="card__hashtag-inner">
+              <span class="card__hashtag-name">
+                ${tags.map((tag) => `<span class="card__hashtag-inner">
+                <span class="card__hashtag-name">
+                  #${tag}
+                </span>
+              </span>`).join(``)}
+              </span>
+            </span>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </article>`.trim();
+};
